fix: move initHandlers into main.go
